Fixed tests, updated readme, added new branch exclusion test
