+++ gcovr/tests/excl-line-branch/reference/clang-10/coverage.txt
