+++ gcovr/tests/excl-line-branch/reference/clang-10/summary_coverage.json
