+++ gcovr/tests/excl-line-branch/reference/gcc-5/coverage.functions.html
