+++ gcovr/tests/excl-line-branch/reference/gcc-5/summary_coverage.json
