+++ gcovr/tests/excl-line-branch/reference/gcc-8/coverage.functions.html
